Update styles in contact us
I changed the style in the contact us page and also moved the inline css
to the customstyle.css with the hours of operation and day of the week
with their own class. You will see a DOW (day of week) class and HoO
(hours of operation) class in the contactus.html as well as the
customstyles.css.